bugfixes to generation of timestamps; need int64_t not int32_t, need 23 chars not 22 or strftime fails, but results in trailing '\0' that breaks sql query? don't quite get it but this seems to work.

+++ JsonParser.h
@@ -10,7 +10,7 @@ using namespace ToolFramework; // for BStore
 
 enum class JsonParserResultType { ints, floats, strings, bools, nulls, stores, empty, undefined };
 struct JsonParserResult {
-	std::vector<int> theints{};
+	std::vector<int64_t> theints{};
 	std::vector<double> thefloats{};
 	std::vector<std::string> thestrings{};
 	std::vector<int> thebools{};
